requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
cycler==0.10.0
discord.py==1.7.3
fuzzywuzzy==0.18.0
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
multidict==5.1.0
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
